+++ src/systems/socket/handledata.rs
@@ -1,6 +1,10 @@
+pub mod handle_entity;
+pub mod handle_general;
+pub mod handle_interface;
+pub mod handle_player;
+pub mod handle_trade;
 pub mod mapper;
 pub mod router;
-pub mod routes;
 
 pub use mapper::{PacketRouter, ServerPackets};
 pub use router::handle_data;
